multiple minor improvements - moved xml functions from utils to replacers - made a prototype to support yaml config file in future - fix output print bugs
