+++ examples/logspout-docker.json
@@ -18,7 +18,7 @@
       "type": "syslog",
       "attrs": {
         "protocol": "udp",
-        "netAddr": "localhost:516",
+        "host": "localhost:516",
         "maxBackups": 3,
         "tag": "logspout"
       }
